Prepare for 2024 open and archive book (README and config)
This is messy because Robert accidentally combined commits that would and would not be cherry-picked later. Take this (squashed) commit as the part that should _not_ be included in 2025, as it includes the altered README and config file to be appropriate for 2024.

This reverts commit e30f31a72591951b9e570ea969fbc0d9444bbb22.

Prepare 2024 branch for making open and archiving (#40)

* update banner config

* Update README.md

* Update README.md

* Update README.md

* change 2025 README to better suit 2024 (archived) book

more updates

deterministic note

typo